+++ rvm/rvmoperatorlsl.c
@@ -48,6 +48,6 @@ void rvm_op_lsl_long(rvmcpu_t *cpu, unsigned short opid, rvmreg_t *res, long op1
 
 void rvm_op_lsl_double(rvmcpu_t *cpu, unsigned short opid, rvmreg_t *res, double op1, double op2)
 {
-	rvm_op_lsl_long(cpu, opid, res, op1, op2);
+	rvm_op_lsl_long(cpu, opid, res, (long)op1, (long)op2);
 }
